requirements.txt
# API and Web Framework
flask==3.0.2
fla***REMOVED***cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0  # To handle API requests

# AI and Machine Learning
anthropic==0.18.1
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.1.post1
boto3==1.28.27  # AWS SDK for Bedrock integration


# Document Processing
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
python-magic==0.4.27
python-docx==1.1.0
python-pptx==0.6.22
PyPDF2==3.0.1

# Text Processing and Embeddings
tiktoken==0.6.0
tokenizers==0.15.2
nltk==3.8.1

# Utils and Type Hints
typing-extensions==4.9.0
pydantic==2.6.1
tenacity==8.2.3
tqdm==4.66.2

# Development Tools
black==24.2.0
flake8==7.0.0
pytest==8.0.2
pytest-cov==4.1.0
